refactor: move worker to a different process
