+++ docker-compose.yml
@@ -16,3 +16,14 @@ services:
     command: bun dev
     volumes:
       - ./src:/app/src
+
+  workers:
+    build: .
+    depends_on:
+      - redis
+    environment:
+      - REDIS_HOST=redis
+      - REDIS_PORT=6379
+    command: bun dev:workers
+    volumes:
+      - ./src:/app/src
