+++ src/app.module.ts
@@ -1,33 +1,8 @@
-import { BullModule } from '@nestjs/bull';
 import { Module } from '@nestjs/common';
-import { ConfigModule, ConfigService } from '@nestjs/config';
-import { TypeOrmModule, type TypeOrmModuleOptions } from '@nestjs/typeorm';
-import { appConfig, type RedisConfig } from './app.config';
+import { SharedModule } from './shared/shared.module';
 import { TransactionsModule } from './transactions/transactions.module';
 
 @Module({
-  imports: [
-    ConfigModule.forRoot({ load: [appConfig], isGlobal: true }),
-    TypeOrmModule.forRootAsync({
-      inject: [ConfigService],
-      useFactory: (configService: ConfigService) => {
-        const typeOrmConfig =
-          configService.get<TypeOrmModuleOptions>('typeOrm');
-        return typeOrmConfig as TypeOrmModuleOptions;
-      },
-    }),
-    BullModule.forRootAsync({
-      inject: [ConfigService],
-      useFactory: (configService: ConfigService) => {
-        const redis = configService.get<RedisConfig>('redis');
-        return {
-          redis,
-        };
-      },
-    }),
-    TransactionsModule,
-  ],
-  controllers: [],
-  providers: [],
+  imports: [SharedModule, TransactionsModule],
 })
 export class AppModule {}
