+++ src/workers/transactions.worker.ts
@@ -1,7 +1,7 @@
 import { Process, Processor } from '@nestjs/bull';
 import { ConflictException, Inject, Logger } from '@nestjs/common';
 import type { Job } from 'bull';
-import type { CreateTransactionDto } from './dto/create-transaction.dto';
+import type { CreateTransactionDto } from '../transactions/dto/create-transaction.dto';
 import { TransactionsRepository } from './transactions.repository';
 
 @Processor('transactions')
@@ -14,15 +14,11 @@ export class TransactionsWorker {
   ) {}
 
   @Process()
-  async process({ id, data }: Job<CreateTransactionDto>) {
+  async process({ data: { userId, amount } }: Job<CreateTransactionDto>) {
     try {
-      const existingTransaction = await this.repository.findById(+id.valueOf());
-      if (existingTransaction) {
-        throw new ConflictException('User has already made a transaction');
-      }
       await this.repository.create({
-        user: data.userId,
-        amount: data.amount,
+        user: userId,
+        amount,
       });
       return 'ok';
     } catch (err) {
